FIXED: Compilation on x86_64 by Martin Capitanio
git-svn-id: http://svn.code.sf.net/p/podofo/code/podofobrowser/trunk@968 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/hexwidget/QHexView.h
@@ -104,7 +104,7 @@ private:
 					// page; for a sequential I/O device it's everything
 					// up to the current point.
 	
-	mutable size_t m_bufOffset;	// Offset in bytes of the buffer into the data.
+	mutable unsigned int m_bufOffset;	// Offset in bytes of the buffer into the data.
 
 public:
 	/**
